+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 2.9.13
 -----------------
 - Always use SchoolInfo() instead of querying SCHOOLS DB table, programwide
 - Unset current student after setting new current school in Schools.php & CopySchool.php
+- Check if Update() version < ROSARIO_VERSION in Update.fnc.php
 
 Changes in 2.9.12
 -----------------